+++ apps/data-api/src/entities/Platform.ts
@@ -1,10 +1,7 @@
 import { Address, BigInt as GraphBigInt, ethereum } from "@graphprotocol/graph-ts";
 import { ensureGlobalSettings } from "../entities/GlobalSettings";
-import { AuctionSettled } from "../generated/ETSAuctionHouse/ETSAuctionHouse";
-import { Transfer } from "../generated/ETSToken/ETSToken";
 import { GlobalSettings, Platform } from "../generated/schema";
-import { ADDED, MODULO, ONE, PAUSED, UNPAUSED, ZERO, ZERO_ADDRESS } from "../utils/constants";
-import { ensureAuction } from "./Auction";
+import { ADDED, ONE, PAUSED, UNPAUSED, ZERO, ZERO_ADDRESS } from "../utils/constants";
 
 export function ensurePlatform(event: ethereum.Event | null): Platform {
   let platform = Platform.load("ETSPlatform");
@@ -18,13 +15,10 @@ export function ensurePlatform(event: ethereum.Event | null): Platform {
     platform.tagsCount = ZERO;
     platform.taggingRecordsCount = ZERO;
     platform.taggingFeesRevenue = ZERO;
-    platform.auctionsSettled = ZERO;
-    platform.auctionRevenue = ZERO;
-    platform.relayerCountActive = ZERO;
-    platform.relayerCountLifetime = ZERO;
+    platform.channelCountActive = ZERO;
+    platform.channelCountLifetime = ZERO;
     platform.taggerCount = ZERO;
     platform.creatorCount = ZERO;
-    platform.ownerCount = ZERO;
     platform.save();
   }
   return platform as Platform;
@@ -55,31 +49,20 @@ export function updatePlatformTaggingRecordStats(tagIds: string[] | null, event:
   updatePlatformRevenue(platform, tagIds, settings);
 }
 
-export function updatePlatformTagStats(event: Transfer): void {
-  const platform = ensurePlatform(event);
-  const fromAddress = event.params.from;
-  const zeroAddress = Address.fromString(ZERO_ADDRESS);
-
-  if (fromAddress.equals(zeroAddress)) {
-    platform.tagsCount = platform.tagsCount.plus(ONE);
-    platform.save();
-  }
-}
-
-export function updateRelayerCount(action: GraphBigInt, event: ethereum.Event): void {
+export function updateChannelCount(action: GraphBigInt, event: ethereum.Event): void {
   const platform = ensurePlatform(event);
 
   if (action === ADDED) {
-    platform.relayerCountLifetime = platform.relayerCountLifetime.plus(ONE);
-    platform.relayerCountActive = platform.relayerCountActive.plus(ONE);
+    platform.channelCountLifetime = platform.channelCountLifetime.plus(ONE);
+    platform.channelCountActive = platform.channelCountActive.plus(ONE);
   }
 
   if (action === PAUSED) {
-    platform.relayerCountActive = platform.relayerCountActive.minus(ONE);
+    platform.channelCountActive = platform.channelCountActive.minus(ONE);
   }
 
   if (action === UNPAUSED) {
-    platform.relayerCountActive = platform.relayerCountActive.plus(ONE);
+    platform.channelCountActive = platform.channelCountActive.plus(ONE);
   }
   platform.save();
 }
@@ -94,24 +77,4 @@ export function updateCreatorCount(event: ethereum.Event): void {
   const platform = ensurePlatform(event);
   platform.creatorCount = platform.creatorCount.plus(ONE);
   platform.save();
-}
-
-export function updateOwnerCount(event: ethereum.Event): void {
-  const platform = ensurePlatform(event);
-  platform.ownerCount = platform.ownerCount.plus(ONE);
-  platform.save();
-}
-
-export function updatePlatformAuctionStats(auctionId: GraphBigInt, event: AuctionSettled): void {
-  const auction = ensureAuction(auctionId, event);
-  //const tag = ensureTag(GraphBigInt.fromString(auction.tag), event);
-  const platform = ensurePlatform(event);
-  if (platform && event) {
-    // pull percentages from settings.
-    const settings = ensureGlobalSettings();
-    const platformAuctionRevenue = auction.amount.times(settings.platformPercentage).div(MODULO);
-    platform.auctionRevenue = platform.auctionRevenue.plus(platformAuctionRevenue);
-    platform.auctionsSettled = platform.auctionsSettled.plus(ONE);
-    platform.save();
-  }
 }