+++ apps/data-api/src/entities/Release.ts
@@ -22,10 +22,6 @@ export function ensureRelease(): Release {
     release.etsTargetVersion = ZERO;
     release.etsTargetVersionDate = ZERO;
 
-    release.etsAuctionHouse = ZERO.toString();
-    release.etsAuctionHouseVersion = ZERO;
-    release.etsAuctionHouseVersionDate = ZERO;
-
     release.save();
   }